Moved NameNamespaceAutosuggest
