--- src/components/namespaceautosuggest/namespaceautosuggest.module.css
@@ -1,6 +1,0 @@
-@import "../../core/styles/index.css";
-
-.entitySetInput {
-    height: 40px;
-    width: 200px;
-}
